Remove unused upload-content module and main.go file; refactor upload-service to use updated config path and improve server port retrieval.
