--- backend/upload-content/go.mod
@@ -1,3 +1,0 @@
-module github.com/jeison-nura/animetest/backend/upload-content
-
-go 1.23.3
