+++ backend/upload-service/main.go
@@ -5,17 +5,17 @@ import (
 	"log"
 	"net/http"
 
-	"github.com/animetest/backend/upload-service/internal/config"
+	"github.com/animetest/backend/upload-service/internal/infrastructure/config"
 	"github.com/gin-gonic/gin"
 )
 
 func main() {
 	cfg := config.LoadConfigFromEnv()
 	router := gin.Default()
-	handler, err := InitializeApp(cfg, router)
+	handler, err := InitializeApp(*cfg, router)
 	if err != nil {
 		log.Fatal(err)
 	}
-	fmt.Printf("Server started on port %s\n", cfg.ServerPort)
-	log.Fatal(http.ListenAndServe(":"+cfg.ServerPort, handler))
+	fmt.Printf("Server started on port %s\n", cfg.GetServerPort())
+	log.Fatal(http.ListenAndServe(":"+cfg.GetServerPort(), handler))
 }
