+++ backend/upload-service/wire.go
@@ -4,13 +4,12 @@
 package main
 
 import (
-	"github.com/animetest/backend/upload-service/internal/config"
+	"github.com/animetest/backend/upload-service/internal/infrastructure/config"
 	"github.com/gin-gonic/gin"
 	"github.com/google/wire"
 )
 
 func InitializeApp(cfg config.Config, router *gin.Engine) (*gin.Engine, error) {
-	wire.Build(
-	)
+	wire.Build()
 	return nil, nil
 }
